[PACA-690] Add idr_dtl_number to mcs details

+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/RdaChange.java
@@ -35,7 +35,7 @@ public class RdaChange<T> {
   @NoArgsConstructor
   public static class Source {
     private String phase;
-    private Integer phaseSeqNum;
+    private Short phaseSeqNum;
     private String transmissionTimestamp;
   }
 }
